requirements.txt
pyelftools>=0.31
pytest>=8.0
